backend/requirements.txt
fastapi>=0.109.2
uvicorn>=0.27.1
python-multipart>=0.0.9
sqlalchemy>=2.0.32
openai>=1.12.0
pymupdf>=1.25.1
python-dotenv>=1.0.1
aiofiles>=23.2.1
python-docx
pandas
openpyxl
